Refactor: move MembershipState to raft_state/
Because `MembershipState` is a sub state of `RaftState`.

And also move ChangeHandler to raft_state/membership_state; Move tests out of mod file.
