+++ openraft/src/lib.rs
@@ -89,7 +89,6 @@ pub use crate::entry::Entry;
 pub use crate::entry::EntryPayload;
 pub use crate::membership::EffectiveMembership;
 pub use crate::membership::Membership;
-pub use crate::membership::MembershipState;
 pub use crate::membership::StoredMembership;
 pub use crate::metrics::RaftMetrics;
 pub use crate::network::RPCTypes;
@@ -101,6 +100,7 @@ pub use crate::node::Node;
 pub use crate::node::NodeId;
 pub use crate::raft::Raft;
 pub use crate::raft::RaftTypeConfig;
+pub use crate::raft_state::MembershipState;
 pub use crate::raft_state::RaftState;
 pub use crate::raft_types::LogId;
 pub use crate::raft_types::LogIdOptionExt;
