+++ openraft/src/membership/mod.rs
@@ -1,8 +1,6 @@
-mod change_handler;
 mod effective_membership;
 mod into_nodes;
 #[allow(clippy::module_inception)] mod membership;
-mod membership_state;
 mod stored_membership;
 
 #[cfg(feature = "bench")]
@@ -10,12 +8,9 @@ mod stored_membership;
 mod bench;
 
 #[cfg(test)] mod effective_membership_test;
-#[cfg(test)] mod membership_state_test;
 #[cfg(test)] mod membership_test;
 
-pub(crate) use change_handler::ChangeHandler;
 pub use effective_membership::EffectiveMembership;
 pub use into_nodes::IntoNodes;
 pub use membership::Membership;
-pub use membership_state::MembershipState;
 pub use stored_membership::StoredMembership;
